ipam: Remove AllocationLimits interface
The allocation limits have been moved into a generic section of the
CilumNode resource so it is no longer required for the specific IPAM
implementation to implement the retrieval of these values.

This means that some ENI specific code will live in pkg/ipam for
backwards compatibility, it can be removed in Cilium 1.9.

Signed-off-by: Thomas Graf <thomas@cilium.io>